backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
greenlet>=3.0.0
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
aiosqlite>=0.22.0
pgvector>=0.2.4
redis>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Task Scheduling
apscheduler>=3.10.4

# HTTP Client & Tools
httpx>=0.26.0
aiofiles>=23.2.1

# Development & Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

# LLM & AI
openai>=1.10.0
sentence-transformers>=2.2.2
langchain-text-splitters>=0.0.1
tiktoken>=0.5.2

# Discovered Dependencies
sse-starlette
loguru
json-repair
typing-extensions
email-validator>=2.0.0
slowapi>=0.1.9
tenacity
prometheus-fastapi-instrumentator

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
networkx>=3.2.1

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
